--- a/lib/handler.ex
+++ b/lib/handler.ex
@@ -6,8 +6,11 @@ defmodule Handler do
   """
 
   alias Handler.{OOM, ProcessExit, Timeout}
-  @type opts :: keyword()
+  @type opts :: list(opt())
+  @type opt :: {:max_ms, milliseconds()} | {:max_heap_bytes, bytes()}
   @type exception :: OOM.t() | ProcessExit.t() | Timeout.t()
+  @type bytes :: non_neg_integer()
+  @type milliseconds :: non_neg_integer()
 
   @doc """
   Run a potentially dangerous function in a safe way.
@@ -28,6 +31,7 @@ defmodule Handler do
 
   """
   def run(fun, opts \\ []) do
+    validate_opts!(opts)
     max_ms = max_ms(opts)
     max_heap_bytes = max_heap_bytes(opts)
 
@@ -97,5 +101,19 @@ defmodule Handler do
       Process.flag(:max_heap_size, bytes_to_words(max_heap_bytes))
       fun.()
     end)
+  end
+
+  defp validate_opts!([]), do: :ok
+  defp validate_opts!(opts) when is_list(opts) do
+    Enum.each(opts, &validate_opt!/1)
+  end
+  defp validate_opts!(_other) do
+    raise ArgumentError, "Invalid opts provided, not a list"
+  end
+
+  defp validate_opt!({:max_ms, number}) when is_integer(number) and number > 0, do: :ok
+  defp validate_opt!({:max_heap_bytes, number}) when is_integer(number) and number > 0, do: :ok
+  defp validate_opt!(other) do
+    raise ArgumentError, "Invalid option #{inspect(other)}"
   end
 end