In #4 @marcelloma caught a typo where I was passing the wrong option name in one of our tests (see https://github.com/SpiffInc/handler/pull/4/files#r793676340)

This PR aims to clean this up as well as providing better typespecs around our options. For instance, the options accepted by `Handler` are slightly different than the ones accepted by `Pool` (ie the `task_name`), so it's helpful to document those as two separate types and I broke out all of the option handling code to its own module as well.