@mmmries 2022-02-10
I also did a quick check on the performance implications of validating options. I ran all of our benchmarks both before and after this change.

## Before


## After

```
michaelries@michaels-mbp handler % mix run bench/overhead_bench.exs 
Operating System: macOS
CPU Information: Apple M1
Number of Available Cores: 8
Available memory: 16 GB
Elixir 1.12.3
Erlang 24.1.4

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 10 s
memory time: 0 ns
parallel: 1
inputs: none specified
Estimated total run time: 48 s

...eliding some output...

Name                     ips        average  deviation         median         99th %
bare function     40630.63 K      0.0246 μs  ±8400.00%           0 μs       0.100 μs
pooled              118.43 K        8.44 μs   ±152.70%           7 μs          24 μs
composed pools       81.62 K       12.25 μs   ±131.47%          10 μs          50 μs
handled               2.64 K      378.44 μs    ±25.22%         369 μs         581 μs

...eliding comparison...

michaelries@michaels-mbp handler % mix run bench/pool_throughput_bench.exs 
Operating System: macOS
CPU Information: Apple M1
Number of Available Cores: 8
Available memory: 16 GB
Elixir 1.12.3
Erlang 24.1.4

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 10 s
memory time: 0 ns
parallel: 16
inputs: none specified
Estimated total run time: 48 s

...eliding some output...

Name                       ips        average  deviation         median         99th %
bare function        8368.34 K       0.119 μs ±11266.16%           0 μs        0.99 μs
pooled function         8.04 K      124.42 μs    ±42.03%      117.99 μs      268.99 μs
composed pool           4.75 K      210.69 μs    ±36.65%      195.99 μs      431.99 μs
handled function        0.66 K     1511.93 μs    ±78.88%     1120.49 μs     5400.99 μs

...eliding comparison...
```